--- a/src/activity_handlers.cpp
+++ b/src/activity_handlers.cpp
@@ -39,6 +39,7 @@
 #include "event_bus.h"
 #include "fault.h"
 #include "field_type.h"
+#include "fstream_utils.h"
 #include "flat_set.h"
 #include "game.h"
 #include "game_constants.h"