--- a/src/advanced_inv.cpp
+++ b/src/advanced_inv.cpp
@@ -51,6 +51,7 @@
 #include "ui_manager.h"
 #include "uistate.h"
 #include "units.h"
+#include "units_utility.h"
 #include "vehicle.h"
 #include "vehicle_selector.h"