--- a/src/auto_pickup.cpp
+++ b/src/auto_pickup.cpp
@@ -12,6 +12,7 @@
 #include "cursesdef.h"
 #include "debug.h"
 #include "filesystem.h"
+#include "fstream_utils.h"
 #include "game.h"
 #include "input.h"
 #include "item.h"