--- a/src/bionics.cpp
+++ b/src/bionics.cpp
@@ -72,6 +72,7 @@
 #include "ui.h"
 #include "ui_manager.h"
 #include "units.h"
+#include "units_utility.h"
 #include "value_ptr.h"
 #include "vehicle.h"
 #include "vpart_position.h"