--- a/src/cata_utility.h
+++ b/src/cata_utility.h
@@ -256,13 +256,6 @@ bool string_starts_with( const std::string &s1, const std::string &s2 );
  */
 bool string_ends_with( const std::string &s1, const std::string &s2 );
 
-/** Used as a default filter in various functions */
-template<typename T>
-bool return_true( const T & )
-{
-    return true;
-}
-
 /**
  * Joins a vector of `std::string`s into a single string with a delimiter/joiner
  */