… src/disease.h
… src/distribution_grid.cpp
… src/distribution_grid.h
… src/editmap.cpp
… src/effect.cpp
… src/filesystem.cpp
… src/filter_utils.h
… src/font_loader.h
… src/fstream_utils.h
… src/game.cpp
… src/game.h
… src/game_inventory.cpp
… src/grab.cpp
… src/handle_liquid.cpp
… src/help.cpp
… src/iexamine.cpp
… src/init.cpp
… src/input.cpp
… src/inventory.h
… src/inventory_ui.cpp
… src/item.cpp
… src/item.h
… src/item_location.h
… src/itype.h
… src/iuse.cpp
… src/iuse_actor.cpp
… src/iuse_software_sokoban.cpp
… src/json.h
… src/language.cpp
… src/magic_enchantment.h