#### Purpose of change
Improving build times, reducing header dependencies

#### Describe the solution
Replaced some `#include`s with forward declarations, broke up some headers into multiple smaller headers / modules, replaced a couple class members with pointers using `pimpl`.

#### Describe alternatives you've considered
Precompiled headers?

#### Testing
Game compiles a bit faster
```
time make -j4

Before:
real    25m18,004s
user    86m4,551s
sys     3m28,834s

After:
real    24m8,300s
user    83m6,683s
sys     3m18,660s
```