@Coolthulhu 2021-03-30
> Precompiled headers?

Precompiled headers are a good idea, but removing false dependencies is worth more than just faster compilation.